+++ app/src/main/java/com/doctor/yumyum/presentation/ui/searchtaste/SearchTasteActivity.kt
@@ -1,4 +1,4 @@
-package com.doctor.yumyum.presentation.ui.search
+package com.doctor.yumyum.presentation.ui.searchtaste
 
 import android.os.Bundle
 import com.doctor.yumyum.R
