+++ app/src/main/res/layout/activity_search_hashtag.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.search.SearchHashtagActivity">
+        tools:context=".presentation.ui.searchhashtag.SearchHashtagActivity">
 
         <ImageButton
             android:id="@+id/search_hashtag_ib_back"
