+++ app/src/main/res/layout/activity_search_taste.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.search.SearchTasteActivity">
+        tools:context=".presentation.ui.searchtaste.SearchTasteActivity">
 
         <include
             android:id="@+id/search_taste_appbar"
